pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "memories"
version = "0.0.1"
authors = [
  { name="Remy Joseph", email="remy.joseph11@gmail.com" },
]
description = "A package to automatically create memory album with synology stationnns."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "psycopg2-binary",
    "pandas",
    "pytest",
    "numpy",
]